% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_masterpull.R
\name{long_masterpull}
\alias{long_masterpull}
\title{A function to import a longitudinal version of EdBuild's master dataset}
\format{A data frame with 111,264 observations and 43 variables. To view
  descriptions of variable names and sources for each use
  \code{master_codebook()}}
\usage{
long_masterpull()
}
\value{
A dataframe where each observation is a school district.
}
\description{
This function imports a longitudinal dataset of EdBuild's master data for the
years 2013- 2018. The master dataset is a compilation of national level
school district data from the US Census Annual Survey of School System
Finances (F33); US Census Small Area Income and Poverty Estimates (SAIPE);
National Center for Education Statistics (NCES) Common Core of Data (CCD);
and Education Demographic and Geographic Estimates (EDGE). Cost adjustments
were calculated using C2ER.
}
\examples{
\donttest{long_master <- long_masterpull()}
}
\seealso{
\code{\link{master_codebook}}, \code{\link{masterpull}}
}
\keyword{CCD}
\keyword{EDGE}
\keyword{EdBuild}
\keyword{F33}
\keyword{SAIPE}
\keyword{data}
\keyword{master}
