% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_codebook.R
\name{master_codebook}
\alias{master_codebook}
\title{A function to view EdBuild's data sources and variable names}
\usage{
master_codebook()
}
\value{
A dataframe where each observation is a variable from the master dataset.
}
\description{
This function imports a codebook for EdBuild's master data. The
codebook tells the user what each variable name represents and the source of
each variable.
}
\examples{
codebook <- master_codebook()
}
\seealso{
\link{long_masterpull}, \code{\link{masterpull}}
}
\keyword{CCD}
\keyword{EDGE}
\keyword{EdBuild}
\keyword{F33}
\keyword{SAIPE}
\keyword{codebook}
\keyword{data}
\keyword{master}
