% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round2.R
\name{round2}
\alias{round2}
\title{A function to round 0.5 down}
\usage{
round2(x, n)
}
\arguments{
\item{x}{a numeric vector}

\item{n}{number of decimal places}
}
\value{
The value returned is a numeric vector, x, rounded to n decimals
  places.
}
\description{
This function rounds values based on mathematical rules. That is, numbers
ending below 0.5 round down and numbers ending with 0.5 and greater round up.
}
\examples{
round2(143.05, 1)
round2(143.048, 2)
}
\references{
\url{https://stackoverflow.com/questions/12688717/round-up-from-5}
}
\keyword{round}
